+++ apps/pie-storybook/stories/pie-input.stories.ts
@@ -1,5 +1,7 @@
 import { html } from 'lit';
 import { ifDefined } from 'lit/directives/if-defined.js';
+import { action } from '@storybook/addon-actions';
+import { useArgs as UseArgs } from '@storybook/preview-api';
 
 /* eslint-disable import/no-duplicates */
 import '@justeattakeaway/pie-input';
@@ -14,6 +16,8 @@ type InputStoryMeta = StoryMeta<InputProps>;
 
 const defaultArgs: InputProps = {
     type: 'text',
+    value: '',
+    name: 'testName',
 };
 
 const inputStoryMeta: InputStoryMeta = {
@@ -28,6 +32,20 @@ const inputStoryMeta: InputStoryMeta = {
                 summary: 'text',
             },
         },
+        value: {
+            description: 'The value of the input (used as a key/value pair in HTML forms with `name`).',
+            control: 'text',
+            defaultValue: {
+                summary: '',
+            },
+        },
+        name: {
+            description: 'The name of the input (used as a key/value pair with `value`). This is required in order to work properly with forms.',
+            control: 'text',
+            defaultValue: {
+                summary: '',
+            },
+        },
     },
     args: defaultArgs,
     parameters: {
@@ -38,53 +56,75 @@ const inputStoryMeta: InputStoryMeta = {
     },
 };
 
-export default inputStoryMeta;
+const Template = ({ type, value, name }: InputProps) => {
+    const [, updateArgs] = UseArgs();
 
-const Template = ({ type }: InputProps) => html`
-    <pie-input type=${ifDefined(type)}></pie-input>
-`;
+    function onInput (event: InputEvent) {
+        const inputElement = event.target as HTMLInputElement;
+        updateArgs({ value: inputElement?.value });
 
-const FormTemplate: TemplateFunction<InputProps> = (props: InputProps) => html`
-<h2>Fake form</h2>
-<form id="testForm">
-        <p>Required fields are followed by <strong><span aria-label="required">*</span></strong>.</p>
-        <section>
-            <h2>Contact information</h2>
-            <p>
-                <label for="name">
-                    <span>Name: </span>
-                    <strong><span aria-label="required">*</span></strong>
-                </label>
-                ${Template({ ...props, type: 'text' })}
-            </p>
-        </section>
-        <section style="display: flex; gap: var(--dt-spacing-a); justify-content: flex-end; flex-wrap: wrap; margin-top: var(--dt-spacing-b);">
-            <pie-button type="reset" variant="secondary">Reset</pie-button>
-            <pie-button type="submit" variant="primary">Submit</pie-button>
-        </section>
-    </form>
-    <p id="formLog" style="display: none; font-size: 2rem; color: var(--dt-color-support-positive);"></p>
-    <script>
-        // Display a success message to the user when they submit the form
-        const form = document.querySelector('#testForm');
-        const formLog = document.querySelector('#formLog');
-        form.addEventListener('submit', (e) => {
-            e.preventDefault();
+        action('input')({
+            data: event.data,
+            value: inputElement.value,
+        });
+    }
 
-            formLog.innerHTML = 'Form submitted!';
-            formLog.style.display = 'block';
+    function onChange (event: CustomEvent) {
+        action('change')({
+            detail: event.detail,
+        });
+    }
 
-            // Reset the success message after roughly 8 seconds
-            setTimeout(() => {
-                formLog.innerHTML = '';
-                formLog.style.display = 'none';
-            }, 8000);
+    return html`
+    <pie-input
+        type="${ifDefined(type)}"
+        .value="${value}"
+        name="${ifDefined(name)}"
+        @input="${onInput}"
+        @change="${onChange}"></pie-input>
+    `;
+};
+
+const FormTemplate: TemplateFunction<InputProps> = (props: InputProps) => {
+    function onSubmit (event: SubmitEvent) {
+        event.preventDefault();
+
+        const form = event.target as HTMLFormElement;
+        const formData = new FormData(form);
+
+        const data: { [key: string]: string } = {};
+        formData.forEach((value, key) => {
+            data[key] = value.toString();
         });
 
-    </script>
-`;
+        action('submit')({
+            ...data,
+        });
+    }
+
+    return html`
+        <h2>Fake form</h2>
+        <form id="testForm" @submit="${onSubmit}">
+            <section>
+                <h2>Contact information</h2>
+                <p>
+                    <label for="name">
+                        <span>Name: </span>
+                    </label>
+                    ${Template({ ...props, type: 'text' })}
+                </p>
+            </section>
+            <section style="display: flex; gap: var(--dt-spacing-a); justify-content: flex-end; flex-wrap: wrap; margin-top: var(--dt-spacing-b);">
+                <pie-button type="reset" variant="secondary">Reset</pie-button>
+                <pie-button type="submit" variant="primary">Submit</pie-button>
+            </section>
+        </form>
+    `;
+};
 
 const createInputStoryWithForm = createStory<InputProps>(FormTemplate, defaultArgs);
 
 export const Default = createStory<InputProps>(Template, defaultArgs)();
 export const FormIntegration = createInputStoryWithForm();
+
+export default inputStoryMeta;