+++ packages/components/pie-input/src/defs.ts
@@ -5,4 +5,14 @@ export interface InputProps {
      * The type of HTML input to render.
      */
     type?: typeof types[number];
+
+    /**
+     * The value of the input (used as a key/value pair in HTML forms with `name`).
+     */
+    value: string;
+
+    /**
+     * The name of the input (used as a key/value pair with `value`). This is required in order to work properly with forms.
+     */
+    name: string;
 }