+++ packages/components/pie-input/src/index.ts
@@ -1,9 +1,12 @@
-import { LitElement, html, unsafeCSS } from 'lit';
+import {
+    LitElement, html, unsafeCSS, PropertyValues,
+} from 'lit';
 import { property } from 'lit/decorators.js';
 import { ifDefined } from 'lit/directives/if-defined.js';
+import { live } from 'lit/directives/live.js';
 
 import {
-    validPropertyValues, RtlMixin, defineCustomElement, FormControlMixin,
+    validPropertyValues, RtlMixin, defineCustomElement, FormControlMixin, wrapNativeEvent,
 } from '@justeattakeaway/pie-webc-core';
 
 import styles from './input.scss?inline';
@@ -16,6 +19,8 @@ const componentSelector = 'pie-input';
 
 /**
  * @tagname pie-input
+ * @event {InputEvent} input - when the input value is changed.
+ * @event {CustomEvent} change - when the input value is changed.
  */
 export class PieInput extends FormControlMixin(RtlMixin(LitElement)) implements InputProps {
     static shadowRootOptions = { ...LitElement.shadowRootOptions, delegatesFocus: true };
@@ -24,11 +29,57 @@ export class PieInput extends FormControlMixin(RtlMixin(LitElement)) implements 
     @validPropertyValues(componentSelector, types, 'text')
     public type: InputProps['type'] = 'text';
 
+    @property({ type: String })
+    public value = '';
+
+    @property({ type: String })
+    public name = '';
+
+    protected firstUpdated (_changedProperties: PropertyValues<this>): void {
+        super.firstUpdated(_changedProperties);
+        this._internals.setFormValue(this.value);
+    }
+
+    protected updated (_changedProperties: PropertyValues<this>): void {
+        super.updated(_changedProperties);
+
+        if (_changedProperties.has('value')) {
+            this._internals.setFormValue(this.value);
+        }
+    }
+
+    /**
+     * Handles data processing in response to the input event. The native input event is left to bubble up.
+     * @param event - The input event.
+     */
+    private handleInput = (event: InputEvent) => {
+        this.value = (event.target as HTMLInputElement).value;
+        this._internals.setFormValue(this.value);
+    };
+
+    /**
+     * Captures the native change event and wraps it in a custom event.
+     * @param event - The change event.
+     */
+    private handleChange = (event: Event) => {
+        // We have to create our own change event because the native one
+        // does not penetrate the shadow boundary.
+
+        // This is because some events set `composed` to `false`.
+        // Reference: https://javascript.info/shadow-dom-events#event-composed
+        const customChangeEvent = wrapNativeEvent(event);
+        this.dispatchEvent(customChangeEvent);
+    };
+
     render () {
-        const { type } = this;
+        const { type, value, name } = this;
 
         return html`<input
             type=${ifDefined(type)}
+            .value=${live(value)}
+            name=${ifDefined(name)}
+            @input=${this.handleInput}
+            @change=${this.handleChange}
             data-test-id="pie-input">`;
     }